test: move server tests
Part of the test clean up.
