+++ test/unit/server/utils/import-resolver.spec.ts
@@ -1,10 +1,10 @@
 import { describe, expect, it } from 'vitest'
-import type { PackageFileTree } from '../../shared/types'
+import type { PackageFileTree } from '../../../../shared/types'
 import {
   createImportResolver,
   flattenFileTree,
   resolveRelativeImport,
-} from '../../server/utils/import-resolver'
+} from '../../../../server/utils/import-resolver'
 
 describe('flattenFileTree', () => {
   it('flattens nested trees into a file set', () => {
